Update coverage2cytosine
Bug fix: If `$genome_folder` is a relative path and `$output_dir` is given, `coverage2cytosine` will fail because it `chdir` to `$output_dir` before `chdir` to `$genome_folder`. To fix this, `coverage2cytosine` will return back to '$parent_dir' before `chdir` to '$genome_folder'. 

The error message is shown as below:

Failed to move to '../../data/ref/bismark_index/'. Make sure the path to the genome folder is correct, and respecify!